Feature : enabled edit and delete action only for question/answer owner; added owner name and update time

+++ resources/views/pages/answer.blade.php
@@ -5,21 +5,28 @@
         <div class="row justify-content-center">
             <div class="col-md-12">
                 <div class="card">
-                    <div class="card-header">Answer</div>
+                    <div class="card-header">Answer
+                        <small class="text-muted float-right">
+                            Updated: {{ $answer->updated_at->diffForHumans() }}
+                        </small>
+                    </div>
 
                     <div class="card-body">
                         <p>{{$answer->body}}</p>
                     </div>
 
                     <div class="card-footer">
-                        <a class="btn btn-success float-right" href="{{ route('answer.edit', ['question_id' => $question, 'answer_id'=>$answer->id,]) }}">
-                            Edit
-                        </a>
-                        {{ Form::open(['method'  => 'DELETE', 'route' => ['answer.destroy', $answer->id, $question]])}}
-                        <button class="btn btn-danger float-right mr-2">
-                            Delete
-                        </button>
-                        {{Form::close()}}
+                        <small class="text-muted float-left">Posted By: {{\App\User::find($answer->user_id)->email}}</small>
+                        @if(Auth::user()->id ==$answer->user_id)
+                            <a class="btn btn-success float-right" href="{{ route('answer.edit', ['question_id' => $question, 'answer_id'=>$answer->id,]) }}">
+                                Edit
+                            </a>
+                            {{ Form::open(['method'  => 'DELETE', 'route' => ['answer.destroy', $answer->id, $question]])}}
+                            <button class="btn btn-danger float-right mr-2">
+                                Delete
+                            </button>
+                            {{Form::close()}}
+                        @endif
                     </div>
                 </div>
             </div>
